# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sudan Export B2B Export Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rfqs:
    get:
      operationId: listRfqs
      tags: [rfqs]
      summary: List RFQs (role-filtered)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: buyerId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of RFQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rfq"
    post:
      operationId: createRfq
      tags: [rfqs]
      summary: Create a new RFQ (buyer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfqInput"
      responses:
        "201":
          description: RFQ created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
  /rfqs/{rfqId}:
    get:
      operationId: getRfq
      tags: [rfqs]
      summary: Get RFQ by ID
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RFQ details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfqDetail"
    patch:
      operationId: updateRfq
      tags: [rfqs]
      summary: Update an RFQ
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfqUpdate"
      responses:
        "200":
          description: RFQ updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
  /rfqs/{rfqId}/status:
    patch:
      operationId: updateRfqStatus
      tags: [rfqs]
      summary: Update RFQ status
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
  /rfqs/{rfqId}/quotations:
    get:
      operationId: listRfqQuotations
      tags: [rfqs]
      summary: List quotations for an RFQ
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotations for this RFQ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quotation"
    post:
      operationId: createQuotation
      tags: [quotations]
      summary: Submit a quotation for an RFQ (supplier)
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationInput"
      responses:
        "201":
          description: Quotation submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"

  # ── Quotations ─────────────────────────────────────────────────────────
  /quotations:
    get:
      operationId: listQuotations
      tags: [quotations]
      summary: List quotations (role-filtered)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: supplierId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of quotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quotation"
  /quotations/{quotationId}:
    get:
      operationId: getQuotation
      tags: [quotations]
      summary: Get quotation by ID
      parameters:
        - name: quotationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationDetail"
  /quotations/{quotationId}/status:
    patch:
      operationId: updateQuotationStatus
      tags: [quotations]
      summary: Award, shortlist, or reject a quotation
      parameters:
        - name: quotationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Quotation status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"

  # ── Messages ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RfqItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        targetPrice:
          type: ["number", "null"]
        specifications:
          type: ["string", "null"]
      required:
        - id
        - productId
        - productName
        - quantity
        - unit
    RfqItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unit:
          type: string
        targetPrice:
          type: ["number", "null"]
        specifications:
          type: ["string", "null"]
      required:
        - productId
        - quantity
        - unit
    Rfq:
      type: object
      properties:
        id:
          type: integer
        referenceNumber:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [draft, submitted, open, quoting, shortlisted, awarded, closed, cancelled]
        buyerId:
          type: string
        buyerName:
          type: ["string", "null"]
        buyerCompanyName:
          type: ["string", "null"]
        destinationCountry:
          type: ["string", "null"]
        deliveryPort:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        incoterms:
          type: ["string", "null"]
        deliveryDeadline:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        quotationCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - referenceNumber
        - title
        - status
        - buyerId
        - quotationCount
        - createdAt
        - updatedAt
    RfqDetail:
      allOf:
        - $ref: "#/components/schemas/Rfq"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/RfqItem"
          required:
            - items
    RfqInput:
      type: object
      properties:
        title:
          type: string
        destinationCountry:
          type: ["string", "null"]
        deliveryPort:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        incoterms:
          type: ["string", "null"]
        deliveryDeadline:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/RfqItemInput"
      required:
        - title
        - items
    RfqUpdate:
      type: object
      properties:
        title:
          type: string
        destinationCountry:
          type: ["string", "null"]
        deliveryPort:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        incoterms:
          type: ["string", "null"]
        deliveryDeadline:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    StatusUpdate:
      type: object
      properties:
        status:
          type: string
        reason:
          type: ["string", "null"]
      required:
        - status

    # ── Quotation ────────────────────────────────────────────────────────
    QuotationItem:
      type: object
      properties:
        id:
          type: integer
        rfqItemId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        totalPrice:
          type: number
        currency:
          type: string
        specifications:
          type: ["string", "null"]
      required:
        - id
        - rfqItemId
        - productName
        - quantity
        - unit
        - unitPrice
        - totalPrice
        - currency
    QuotationItemInput:
      type: object
      properties:
        rfqItemId:
          type: integer
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        currency:
          type: string
        specifications:
          type: ["string", "null"]
      required:
        - rfqItemId
        - quantity
        - unit
        - unitPrice
        - currency
    Quotation:
      type: object
      properties:
        id:
          type: integer
        rfqId:
          type: integer
        rfqTitle:
          type: ["string", "null"]
        supplierId:
          type: string
        supplierName:
          type: ["string", "null"]
        supplierCompanyName:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, submitted, shortlisted, awarded, rejected]
        totalAmount:
          type: ["number", "null"]
        currency:
          type: string
        validUntil:
          type: ["string", "null"]
        deliveryTime:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - rfqId
        - supplierId
        - status
        - currency
        - createdAt
        - updatedAt
    QuotationDetail:
      allOf:
        - $ref: "#/components/schemas/Quotation"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/QuotationItem"
          required:
            - items
    QuotationInput:
      type: object
      properties:
        currency:
          type: string
        validUntil:
          type: ["string", "null"]
        deliveryTime:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationItemInput"
      required:
        - currency
        - items

    # ── Message ──────────────────────────────────────────────────────────
...