# lib/api-zod/src/generated/types/rfqUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sudan Export B2B Export Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RfqUpdate {
  title?: string;
  /** @nullable */
  destinationCountry?: string | null;
  /** @nullable */
  deliveryPort?: string | null;
  /** @nullable */
  paymentTerms?: string | null;
  /** @nullable */
  incoterms?: string | null;
  /** @nullable */
  deliveryDeadline?: string | null;
  /** @nullable */
  validUntil?: string | null;
  /** @nullable */
  notes?: string | null;
}